Extract config management and test-libraries cloning into their own modules (#1311)
