+++ quickstart.py
@@ -13,7 +13,6 @@ import shlex
 import socket
 import subprocess
 import sys
-import tempfile
 import threading
 import time
 import uuid
@@ -53,21 +52,27 @@ from flask_session import Session
 from modules import validations, output, persistence, helpers, database, logscan, importer, path_validation, url_validation
 from modules.background_jobs import (
     JOB_TARGET_PAGES,
-    create_background_job as _create_background_job,
+    create_background_job as _create_background_job,  # noqa: F401 (used directly by tests as qs_module._create_background_job)
     get_background_job as _get_background_job,
     get_active_background_job as _get_active_background_job,
     get_active_background_jobs as _get_active_background_jobs,
     update_background_job as _update_background_job,
     clear_active_background_job as _clear_active_background_job,
     ensure_background_job as _ensure_background_job,
-    complete_background_job as _complete_background_job,
-    fail_background_job as _fail_background_job,
+    complete_background_job as _complete_background_job,  # noqa: F401 (used directly by tests as qs_module._complete_background_job)
 )
 from blueprints.validation_routes import bp as validation_routes_bp, refresh_plex_libraries
 from blueprints.asset_routes import bp as asset_routes_bp
 from blueprints.kometa_updates import bp as kometa_updates_bp
 from blueprints.imagemaid_updates import bp as imagemaid_updates_bp
+from blueprints.config_routes import bp as config_routes_bp
+from blueprints.test_libraries_routes import bp as test_libraries_routes_bp
 from modules.assets import build_preview_image_data as _build_preview_image_data, list_overlay_fonts
+from modules.test_libraries import (
+    resolve_test_libraries_paths as _resolve_test_libraries_paths,  # noqa: F401 (used directly by tests as qs_module._resolve_test_libraries_paths)
+    paths_overlap as _paths_overlap,  # noqa: F401 (used directly by tests as qs_module._paths_overlap)
+    safe_to_replace_test_libraries as _safe_to_replace_test_libraries,  # noqa: F401 (used directly by tests as qs_module._safe_to_replace_test_libraries)
+)
 from modules.kometa_install import (
     KOMETA_INSTALL_MODE_MANAGED,
     KOMETA_INSTALL_MODE_EXTERNAL,
@@ -2767,133 +2772,6 @@ def _build_workspace_status_context(config_name, template_list, available_config
     }
 
 
-def _sanitize_config_name(raw_name: str | None) -> str:
-    if not isinstance(raw_name, str):
-        return ""
-    return re.sub(r"[^a-z0-9_]", "", raw_name.strip().lower())
-
-
-def _normalize_config_filename(config_name: str | None) -> str:
-    name = (config_name or "").strip().lower().replace(" ", "_")
-    return name or "default"
-
-
-def _rename_config_files(old_name: str, new_name: str, dry_run: bool = False) -> dict:
-    result = {"success": False, "renamed": [], "skipped": [], "errors": [], "rollback_errors": []}
-    old_norm = _normalize_config_filename(old_name)
-    new_norm = _normalize_config_filename(new_name)
-    if old_norm == new_norm:
-        result["skipped"].append("Normalized filenames are identical.")
-        return result
-
-    config_dir = Path(helpers.CONFIG_DIR)
-    kometa_root = Path(app.config.get("KOMETA_ROOT", "."))
-    config_file = config_dir / f"{old_norm}_config.yml"
-    new_config_file = config_dir / f"{new_norm}_config.yml"
-    kometa_file = kometa_root / "config" / f"{old_norm}_config.yml"
-    new_kometa_file = kometa_root / "config" / f"{new_norm}_config.yml"
-
-    if new_config_file.exists() or new_kometa_file.exists():
-        result["errors"].append("Target config filename already exists.")
-        return result
-
-    archive_root = config_dir / "archives"
-    old_archive = archive_root / old_norm
-    new_archive = archive_root / new_norm
-    old_managed_root = helpers.get_managed_config_artifact_root(old_norm)
-    new_managed_root = helpers.get_managed_config_artifact_root(new_norm)
-    old_managed_dirs = helpers.get_managed_library_artifact_paths(old_norm)
-    new_managed_dirs = helpers.get_managed_library_artifact_paths(new_norm)
-    old_legacy_managed_dirs = helpers.get_legacy_managed_library_artifact_paths(old_norm)
-    new_legacy_managed_dirs = helpers.get_legacy_managed_library_artifact_paths(new_norm)
-    if old_archive.exists():
-        if new_archive.exists():
-            result["errors"].append("Target archive directory already exists.")
-            return result
-        existing_names = {p.name for p in old_archive.glob("*.yml")}
-        for path in old_archive.glob(f"{old_norm}_config_*.yml"):
-            target_name = path.name.replace(f"{old_norm}_config_", f"{new_norm}_config_", 1)
-            if target_name in existing_names and target_name != path.name:
-                result["errors"].append(f"Archive file already exists: {target_name}")
-                return result
-
-    if old_managed_root.exists() and new_managed_root.exists():
-        result["errors"].append(f"Target managed config directory already exists: {new_managed_root}")
-        return result
-
-    for old_managed, new_managed in zip(old_managed_dirs, new_managed_dirs):
-        if old_managed.exists() and new_managed.exists():
-            result["errors"].append(f"Target managed library directory already exists: {new_managed}")
-            return result
-    for old_managed, new_managed in zip(old_legacy_managed_dirs, new_managed_dirs):
-        if old_managed.exists() and new_managed.exists():
-            result["errors"].append(f"Target managed library directory already exists: {new_managed}")
-            return result
-    for old_managed, new_managed in zip(old_legacy_managed_dirs, new_legacy_managed_dirs):
-        if old_managed.exists() and new_managed.exists():
-            result["errors"].append(f"Target legacy managed library directory already exists: {new_managed}")
-            return result
-
-    if dry_run:
-        result["success"] = True
-        return result
-
-    completed = []
-    try:
-        if config_file.exists():
-            config_file.rename(new_config_file)
-            completed.append((config_file, new_config_file))
-            result["renamed"].append(str(new_config_file))
-        if kometa_file.exists():
-            kometa_file.rename(new_kometa_file)
-            completed.append((kometa_file, new_kometa_file))
-            result["renamed"].append(str(new_kometa_file))
-        if old_managed_root.exists():
-            new_managed_root.parent.mkdir(parents=True, exist_ok=True)
-            old_managed_root.rename(new_managed_root)
-            completed.append((old_managed_root, new_managed_root))
-            result["renamed"].append(str(new_managed_root))
-        for old_managed, new_managed in zip(old_managed_dirs, new_managed_dirs):
-            if old_managed.exists():
-                new_managed.parent.mkdir(parents=True, exist_ok=True)
-                old_managed.rename(new_managed)
-                completed.append((old_managed, new_managed))
-                result["renamed"].append(str(new_managed))
-        for old_managed, new_managed in zip(old_legacy_managed_dirs, new_managed_dirs):
-            if old_managed.exists():
-                new_managed.parent.mkdir(parents=True, exist_ok=True)
-                old_managed.rename(new_managed)
-                completed.append((old_managed, new_managed))
-                result["renamed"].append(str(new_managed))
-
-        if old_archive.exists():
-            old_archive.rename(new_archive)
-            completed.append((old_archive, new_archive))
-            result["renamed"].append(str(new_archive))
-            file_ops = []
-            for path in new_archive.glob(f"{old_norm}_config_*.yml"):
-                new_path = new_archive / path.name.replace(f"{old_norm}_config_", f"{new_norm}_config_", 1)
-                if new_path.exists() and new_path != path:
-                    raise FileExistsError(f"Archive file already exists: {new_path.name}")
-                file_ops.append((path, new_path))
-            for src, dst in file_ops:
-                src.rename(dst)
-                completed.append((src, dst))
-                result["renamed"].append(str(dst))
-    except Exception as exc:
-        result["errors"].append(f"Rename failed: {exc}")
-        for src, dst in reversed(completed):
-            try:
-                if Path(dst).exists():
-                    Path(dst).rename(src)
-            except Exception as rollback_exc:
-                result["rollback_errors"].append(str(rollback_exc))
-        return result
-
-    result["success"] = True
-    return result
-
-
 DOTENV = os.path.relpath(os.path.join(helpers.CONFIG_DIR, ".env"))
 load_dotenv(DOTENV, override=True)
 
@@ -2913,6 +2791,8 @@ app.register_blueprint(validation_routes_bp)
 app.register_blueprint(asset_routes_bp)
 app.register_blueprint(kometa_updates_bp)
 app.register_blueprint(imagemaid_updates_bp)
+app.register_blueprint(config_routes_bp)
+app.register_blueprint(test_libraries_routes_bp)
 
 # Run version check at startup
 app.config["VERSION_CHECK"] = helpers.check_for_update()
@@ -3168,259 +3048,6 @@ helpers.ts_log(f"Running on port: {port} | Debug Mode: {'Enabled' if debug_mode 
 @app.route("/")
 def start():
     return redirect(url_for("step", name="001-start"))
-
-
-@app.route("/clear_session", methods=["POST"])
-def clear_session():
-    data = request.values
-    try:
-        config_name = data["name"]
-        if config_name != session["config_name"]:
-            session["config_name"] = config_name
-    except KeyError:  # Handle missing `name` key safely
-        config_name = session.get("config_name")
-
-    persistence.flush_session_storage(config_name)
-
-    # Send message to toast
-    return jsonify(
-        {
-            "status": "success",
-            "message": f"Session storage cleared for '{config_name}'.",
-        }
-    )
-
-
-@app.route("/clear_data/<name>/<section>")
-def clear_data_section(name, section):
-    database.reset_data(name, section)
-    flash("SQLite storage cleared successfully.", "success")
-    return redirect(url_for("start"))
-
-
-@app.route("/clear_data/<name>")
-def clear_data(name):
-    database.reset_data(name)
-    cleanup = helpers.delete_config_artifacts(name, kometa_root=app.config.get("KOMETA_ROOT", "."))
-    for msg in cleanup.get("errors", []):
-        helpers.ts_log(msg, level="WARNING")
-    flash("SQLite storage cleared successfully.", "success")
-    return redirect(url_for("start"))
-
-
-@app.route("/switch-config", methods=["POST"])
-def switch_config():
-    data = request.get_json(silent=True) or {}
-    name = (data.get("name") or "").strip()
-    if not name:
-        return jsonify(success=False, message="Config name is required."), 400
-
-    available = database.get_unique_config_names() or []
-    if name not in available:
-        return jsonify(success=False, message="Config not found."), 404
-
-    session["config_name"] = name
-    try:
-        menu_templates = helpers.get_menu_list()
-        workspace_status = _build_workspace_status_context(name, menu_templates, available_configs=available)
-    except Exception:
-        workspace_status = {}
-    return jsonify(success=True, name=name, workspace_status=workspace_status)
-
-
-@app.route("/activate-config", methods=["POST"])
-def activate_config():
-    data = request.get_json(silent=True) or {}
-    requested_name = data.get("name")
-    name = _sanitize_config_name(requested_name)
-    if not name:
-        return jsonify(success=False, message="Config name is required."), 400
-
-    available = database.get_unique_config_names() or []
-    created = name not in available
-
-    session["config_name"] = name
-
-    if created:
-        seed_payload = {
-            "start": {"config_name": name},
-            "validated_at": utc_now_iso(),
-        }
-        database.save_section_data(
-            name=name,
-            section="start",
-            validated=True,
-            user_entered=True,
-            data=seed_payload,
-        )
-
-    return jsonify(success=True, name=name, created=created)
-
-
-@app.route("/bulk-delete-configs", methods=["POST"])
-def bulk_delete_configs():
-    data = request.get_json(silent=True) or {}
-    names = data.get("names") or []
-    if not isinstance(names, list):
-        return jsonify(success=False, message="Invalid request payload."), 400
-
-    cleaned = [n.strip() for n in names if isinstance(n, str) and n.strip()]
-    if not cleaned:
-        return jsonify(success=False, message="No configs selected."), 400
-
-    available = set(database.get_unique_config_names() or [])
-    deleted = []
-    for name in cleaned:
-        if name in available:
-            database.reset_data(name)
-            cleanup = helpers.delete_config_artifacts(name, kometa_root=app.config.get("KOMETA_ROOT", "."))
-            for msg in cleanup.get("errors", []):
-                helpers.ts_log(msg, level="WARNING")
-            deleted.append(name)
-
-    remaining = database.get_unique_config_names() or []
-    current = session.get("config_name")
-    if current in deleted:
-        session["config_name"] = remaining[0] if remaining else namesgenerator.get_random_name()
-        current = session["config_name"]
-
-    return jsonify(success=True, deleted=deleted, remaining=remaining, current=current)
-
-
-@app.route("/orphaned-config-artifacts", methods=["GET"])
-def orphaned_config_artifacts():
-    result = helpers.list_orphaned_config_artifacts(kometa_root=app.config.get("KOMETA_ROOT", "."))
-    status_code = 200 if not result.get("errors") else 500
-    return jsonify(success=not bool(result.get("errors")), orphans=result.get("orphans", []), errors=result.get("errors", [])), status_code
-
-
-@app.route("/orphaned-config-artifacts/versions", methods=["GET"])
-def orphaned_config_artifact_versions():
-    name = request.args.get("name")
-    normalized = helpers.normalize_config_name_for_storage(name)
-    inventory = helpers.list_orphaned_config_artifacts(kometa_root=app.config.get("KOMETA_ROOT", "."))
-    if inventory.get("errors"):
-        return jsonify(success=False, message="Unable to inspect config storage.", errors=inventory["errors"]), 500
-
-    orphan_names = {item.get("name") for item in inventory.get("orphans", []) if item.get("name")}
-    if normalized not in orphan_names:
-        return jsonify(success=False, message="Config bundle is not currently orphaned."), 404
-
-    result = helpers.list_orphaned_config_versions(normalized)
-    return jsonify(success=True, name=normalized, versions=result.get("versions", []))
-
-
-@app.route("/orphaned-config-artifacts/restore", methods=["POST"])
-def restore_orphaned_config_artifact():
-    data = request.get_json(silent=True) or {}
-    name = helpers.normalize_config_name_for_storage(data.get("name"))
-    selected_path = data.get("path")
-    if not name or not isinstance(selected_path, str) or not selected_path.strip():
-        return jsonify(success=False, message="Config bundle name and version path are required."), 400
-
-    available = database.get_unique_config_names() or []
-    if any(existing.lower() == name.lower() for existing in available):
-        return jsonify(success=False, message="Config already exists in the database."), 400
-
-    inventory = helpers.list_orphaned_config_artifacts(kometa_root=app.config.get("KOMETA_ROOT", "."))
-    if inventory.get("errors"):
-        return jsonify(success=False, message="Unable to inspect config storage.", errors=inventory["errors"]), 500
-
-    orphan_names = {item.get("name") for item in inventory.get("orphans", []) if item.get("name")}
-    if name not in orphan_names:
-        return jsonify(success=False, message="Config bundle is not currently orphaned."), 404
-
-    versions = helpers.list_orphaned_config_versions(name).get("versions", [])
-    version_lookup = {entry.get("path"): entry for entry in versions if entry.get("path")}
-    if selected_path not in version_lookup:
-        return jsonify(success=False, message="Selected version is not available for restore."), 400
-
-    try:
-        source_path = Path(selected_path).resolve()
-        yaml_text = source_path.read_text(encoding="utf-8", errors="replace")
-    except Exception as exc:
-        return jsonify(success=False, message=f"Unable to read the selected config version: {exc}"), 400
-
-    config_data = importer.load_yaml_config(yaml_text)
-    if not config_data:
-        return jsonify(success=False, message="Selected config version could not be parsed as YAML."), 400
-
-    payload, report = importer.prepare_import_payload(config_data, set(), set())
-    if not payload:
-        return jsonify(success=False, message="Selected config version has no importable Quickstart sections."), 400
-
-    for section, data_blob in payload.items():
-        database.save_section_data(
-            name=name,
-            section=section,
-            validated=False,
-            user_entered=True,
-            data=data_blob,
-        )
-
-    config_dir = Path(helpers.CONFIG_DIR)
-    current_file = (config_dir / f"{name}_config.yml").resolve()
-    if source_path != current_file:
-        helpers.save_to_named_config(yaml_text, name)
-    else:
-        kometa_config_dir = Path(app.config.get("KOMETA_ROOT", ".")) / "config"
-        try:
-            kometa_config_dir.mkdir(parents=True, exist_ok=True)
-            (kometa_config_dir / current_file.name).write_text(yaml_text, encoding="utf-8")
-        except OSError as exc:
-            helpers.ts_log(f"Failed to sync restored config to Kometa: {exc}", level="WARNING")
-
-    session["config_name"] = name
-    try:
-        menu_templates = helpers.get_menu_list()
-        workspace_status = _build_workspace_status_context(name, menu_templates, available_configs=database.get_unique_config_names() or [])
-    except Exception:
-        workspace_status = {}
-
-    return jsonify(
-        success=True,
-        config_name=name,
-        restored_path=str(source_path),
-        imported_sections=sorted(payload.keys()),
-        report_summary=report.summary(),
-        workspace_status=workspace_status,
-    )
-
-
-@app.route("/orphaned-config-artifacts/delete", methods=["POST"])
-def delete_orphaned_config_artifacts():
-    data = request.get_json(silent=True) or {}
-    names = data.get("names") or []
-    if not isinstance(names, list):
-        return jsonify(success=False, message="Invalid request payload."), 400
-
-    selected = [helpers.normalize_config_name_for_storage(name) for name in names if str(name or "").strip()]
-    if not selected:
-        return jsonify(success=False, message="No orphaned configs selected."), 400
-
-    inventory = helpers.list_orphaned_config_artifacts(kometa_root=app.config.get("KOMETA_ROOT", "."))
-    if inventory.get("errors"):
-        return jsonify(success=False, message="Unable to inspect config storage.", errors=inventory["errors"]), 500
-
-    orphan_bundles = {item.get("name"): item for item in inventory.get("orphans", []) if isinstance(item, dict) and item.get("name")}
-    orphan_names = set(orphan_bundles)
-    invalid = [name for name in selected if name not in orphan_names]
-    if invalid:
-        return jsonify(success=False, message="Only orphaned config bundles can be deleted here.", invalid=invalid), 400
-
-    deleted = []
-    errors = []
-    for name in selected:
-        result = helpers.delete_orphaned_artifact_bundle(orphan_bundles.get(name))
-        if result.get("errors"):
-            errors.extend(result["errors"])
-        else:
-            deleted.append(name)
-
-    if errors:
-        return jsonify(success=False, deleted=deleted, errors=errors, message="Some orphaned config bundles could not be deleted."), 500
-
-    return jsonify(success=True, deleted=deleted)
 
 
 def _build_logscan_resolution_context(log_dir=None, include_candidate_files=True):
@@ -3940,55 +3567,6 @@ def logscan_trends_log_compress():
         response["compressed_file"] = bool(compressed[0].get("compressed_file"))
         response["compressed_path"] = compressed[0].get("compressed_path")
     return jsonify(response)
-
-
-@app.route("/rename-config", methods=["POST"])
-def rename_config():
-    data = request.get_json(silent=True) or {}
-    old_name = str(data.get("old_name", "")).strip()
-    new_name = _sanitize_config_name(data.get("new_name"))
-    if not old_name or not new_name:
-        return jsonify(success=False, message="Config names are required."), 400
-    if old_name == new_name:
-        return jsonify(success=False, message="New name must be different."), 400
-
-    available = database.get_unique_config_names() or []
-    if old_name not in available:
-        return jsonify(success=False, message="Config not found."), 404
-
-    for name in available:
-        if name.lower() == new_name.lower() and name != old_name:
-            return jsonify(success=False, message="Config name already exists."), 400
-
-    file_check = _rename_config_files(old_name, new_name, dry_run=True)
-    if not file_check.get("success"):
-        return jsonify(success=False, message="Config files are not safe to rename.", details=file_check), 400
-
-    file_result = _rename_config_files(old_name, new_name)
-    if not file_result.get("success"):
-        return jsonify(success=False, message="Failed to rename config files.", details=file_result), 500
-
-    try:
-        update_result = database.rename_config(old_name, new_name)
-    except Exception as exc:
-        helpers.ts_log(f"Failed to update database during rename: {exc}", level="ERROR")
-        rollback = _rename_config_files(new_name, old_name)
-        response = {"success": False, "message": "Failed to update database."}
-        if app.config["QS_DEBUG"]:
-            response["details"] = rollback
-        return jsonify(response), 500
-
-    if not update_result.get("success"):
-        rollback = _rename_config_files(new_name, old_name)
-        response = {"success": False, "message": "Failed to update database."}
-        if app.config["QS_DEBUG"]:
-            response["details"] = rollback
-        return jsonify(response), 500
-
-    if session.get("config_name") == old_name:
-        session["config_name"] = new_name
-
-    return jsonify(success=True, old_name=old_name, new_name=new_name, files=file_result)
 
 
 def count_annotated_lines(text: str) -> dict:
@@ -13587,468 +13165,6 @@ def imagemaid_status():
         )
 
 
-def _normalize_test_libraries_path(raw_path, base_dir):
-    value = str(raw_path or "").strip().strip('"').strip("'")
-    if not value:
-        return ""
-    value = os.path.expandvars(value)
-    value = os.path.expanduser(value)
-    if not os.path.isabs(value):
-        value = os.path.abspath(os.path.join(base_dir, value))
-    return os.path.abspath(value)
-
-
-def _resolve_test_libraries_paths(quickstart_root):
-    base_config_dir = os.path.dirname(sys.executable) if getattr(sys, "frozen", False) else quickstart_root
-    default_final = os.path.join(base_config_dir, "config", "plex_test_libraries")
-    default_tmp = os.path.join(base_config_dir, "config", "tmp")
-    raw_final = app.config.get("QS_TEST_LIBS_PATH") or os.getenv("QS_TEST_LIBS_PATH") or default_final
-    raw_tmp = app.config.get("QS_TEST_LIBS_TMP") or os.getenv("QS_TEST_LIBS_TMP") or default_tmp
-    final_path = _normalize_test_libraries_path(raw_final, base_config_dir) or os.path.abspath(default_final)
-    tmp_path = _normalize_test_libraries_path(raw_tmp, base_config_dir) or os.path.abspath(default_tmp)
-    return base_config_dir, final_path, tmp_path, default_final, default_tmp
-
-
-def _test_libraries_present(path):
-    if not path or not os.path.isdir(path):
-        return False
-    expected_dirs = [
-        os.path.join(path, "test_tv_lib"),
-        os.path.join(path, "test_movie_lib"),
-    ]
-    marker = os.path.join(path, ".test_libraries_version")
-    return all(os.path.isdir(p) for p in expected_dirs) or os.path.exists(marker)
-
-
-def _paths_overlap(path_a, path_b):
-    if not path_a or not path_b:
-        return False
-    try:
-        common = os.path.commonpath([os.path.abspath(path_a), os.path.abspath(path_b)])
-    except ValueError:
-        return False
-    return common == os.path.abspath(path_a) or common == os.path.abspath(path_b)
-
-
-def _ensure_rw_dir(path):
-    if not path:
-        return False, "Path is empty."
-    if os.path.exists(path) and not os.path.isdir(path):
-        return False, "Path exists but is not a directory."
-    try:
-        os.makedirs(path, exist_ok=True)
-    except Exception as e:
-        helpers.ts_log(f"Unable to create folder '{path}': {e}", level="ERROR")
-        return False, "Unable to create folder."
-    test_file = os.path.join(path, f".qs_write_test_{uuid.uuid4().hex}")
-    try:
-        with open(test_file, "w", encoding="utf-8") as f:
-            f.write("test")
-        os.remove(test_file)
-    except Exception as e:
-        helpers.ts_log(f"Unable to write to folder '{path}': {e}", level="ERROR")
-        return False, "Unable to write to folder."
-    return True, ""
-
-
-def _safe_to_replace_test_libraries(path):
-    if not path:
-        return False
-    if not os.path.exists(path):
-        return True
-    if _test_libraries_present(path):
-        return True
-    if os.path.isdir(path) and not os.listdir(path):
-        return True
-    return False
-
-
-@app.route("/check-test-libraries", methods=["POST"])
-def check_test_libraries():
-    data = request.get_json(silent=True) or {}
-    quickstart_root = data.get("quickstart_root", "")
-    if not quickstart_root:
-        return jsonify(success=False, message="Quickstart root path not provided.")
-
-    _, target_path, _, _, _ = _resolve_test_libraries_paths(quickstart_root)
-    resolved_path = os.path.abspath(target_path)
-
-    found = _test_libraries_present(target_path)
-    target_exists = os.path.exists(target_path)
-    unrecognized = bool(target_exists and not found)
-
-    local_sha = ""
-    remote_sha = ""
-    is_outdated = False
-
-    if found:
-        sha_path = os.path.join(target_path, ".test_libraries_version")
-        if os.path.exists(sha_path):
-            try:
-                with open(sha_path, "r") as f:
-                    local_sha = f.read().strip()
-            except Exception:
-                local_sha = ""
-            try:
-                commit_info = requests.get(
-                    "https://api.github.com/repos/chazlarson/plex-test-libraries/commits/main",
-                    timeout=5,
-                ).json()
-                remote_sha = commit_info.get("sha", "")[:7]
-            except Exception:
-                remote_sha = ""
-            if local_sha and remote_sha and local_sha != remote_sha:
-                is_outdated = True
-
-    return jsonify(
-        {
-            "found": bool(found),
-            "target_path": resolved_path,
-            "is_outdated": is_outdated,
-            "local_sha": local_sha,
-            "remote_sha": remote_sha,
-            "target_exists": target_exists,
-            "unrecognized": unrecognized,
-        }
-    )
-
-
-@app.route("/test-libraries-settings", methods=["POST"])
-def update_test_libraries_settings():
-    data = request.get_json(silent=True) or {}
-    quickstart_root = data.get("quickstart_root", "")
-    if not quickstart_root:
-        return jsonify(success=False, message="Quickstart root path not provided.")
-
-    temp_raw = data.get("temp_path", "")
-    final_raw = data.get("final_path", "")
-    confirm = helpers.booler(str(data.get("confirm", "")))
-
-    path_errors = path_validation.validate_payload(
-        {
-            "temp_path": temp_raw,
-            "final_path": final_raw,
-        }
-    )
-    if path_errors:
-        return jsonify(success=False, message="Invalid path values: " + " ".join(path_errors)), 400
-
-    base_config_dir, _, _, default_final, default_tmp = _resolve_test_libraries_paths(quickstart_root)
-    temp_path = _normalize_test_libraries_path(temp_raw or default_tmp, base_config_dir)
-    final_path = _normalize_test_libraries_path(final_raw or default_final, base_config_dir)
-
-    if not temp_path or not final_path:
-        return jsonify(success=False, message="Temp and final paths are required."), 400
-
-    if _paths_overlap(temp_path, final_path):
-        return jsonify(success=False, message="Temp and final paths must be different and cannot be nested."), 400
-
-    ok, msg = _ensure_rw_dir(temp_path)
-    if not ok:
-        return jsonify(success=False, message=msg), 400
-    ok, msg = _ensure_rw_dir(final_path)
-    if not ok:
-        return jsonify(success=False, message=msg), 400
-
-    old_final = _normalize_test_libraries_path(app.config.get("QS_TEST_LIBS_PATH") or default_final, base_config_dir)
-    old_has_libs = _test_libraries_present(old_final)
-    if old_final and final_path != old_final and old_has_libs and not confirm:
-        return (
-            jsonify(
-                success=False,
-                needs_confirm=True,
-                message="Test libraries exist at the previous configured path. Quickstart will not move them.",
-            ),
-            409,
-        )
-
-    final_has_content = False
-    if os.path.isdir(final_path):
-        try:
-            final_has_content = any(os.scandir(final_path))
-        except Exception:
-            final_has_content = True
-    final_is_test_libs = _test_libraries_present(final_path)
-    if final_has_content and not final_is_test_libs and not confirm:
-        return (
-            jsonify(
-                success=False,
-                needs_confirm=True,
-                message="The final path is not empty and does not look like test libraries. Quickstart will replace this folder during install/update.",
-            ),
-            409,
-        )
-
-    helpers.update_env_variable("QS_TEST_LIBS_TMP", temp_path)
-    helpers.update_env_variable("QS_TEST_LIBS_PATH", final_path)
-    os.environ["QS_TEST_LIBS_TMP"] = temp_path
-    os.environ["QS_TEST_LIBS_PATH"] = final_path
-    app.config["QS_TEST_LIBS_TMP"] = temp_path
-    app.config["QS_TEST_LIBS_PATH"] = final_path
-
-    return jsonify(
-        success=True,
-        message="Test library paths saved.",
-        temp_path=temp_path,
-        final_path=final_path,
-        old_path=old_final if old_final and final_path != old_final else "",
-    )
-
-
-@app.route("/clone-test-libraries-start", methods=["POST"])
-def clone_test_libraries_start():
-    """
-    Starts a background job to download and install plex_test_libraries.
-
-    Progress payload shapes by phase:
-      download: {"phase":"download","pct":<int|None>,"text":str,"downloaded":int,"total":int}
-      extract : {"phase":"extract","pct":int,"text":str,"files_done":int,"files_total":int}
-      finalize: {"phase":"finalize","pct":int,"text":str}
-      done    : {"phase":"done","pct":100,"text":str,"target_path":str}
-      error   : {"phase":"error","pct":0,"text":str}
-    """
-    data = request.get_json(silent=True) or {}
-    quickstart_root = data.get("quickstart_root", "")
-
-    if not quickstart_root:
-        return jsonify(success=False, message="Quickstart root path not provided.")
-
-    _, target_path, tmp_root, _, _ = _resolve_test_libraries_paths(quickstart_root)
-    resolved_path = os.path.abspath(target_path)
-    if _paths_overlap(tmp_root, target_path):
-        return jsonify(success=False, message="Temp and final paths must be different and cannot be nested.")
-    if not _safe_to_replace_test_libraries(target_path):
-        return jsonify(
-            success=False,
-            message="Target path exists but does not look like test libraries. Choose an empty folder or one containing test libraries.",
-        )
-
-    active = _get_active_background_job("test_library_install")
-    if active:
-        phase = active.get("phase")
-        if phase and phase not in ["done", "error"]:
-            return jsonify(success=True, job_id=active.get("job_id"), existing_job=True, started_at=active.get("started_epoch"))
-        _clear_active_background_job("test_library_install", job_id=active.get("job_id"))
-    job = _create_background_job(
-        "test_library_install",
-        trigger="manual",
-        phase="queued",
-        status="running",
-        target_page=JOB_TARGET_PAGES.get("test_library_install"),
-        pct=0,
-        text="Queued...",
-        started_epoch=time.time(),
-    )
-    job_id = job["job_id"]
-
-    def worker():
-        zip_url = "https://github.com/chazlarson/plex-test-libraries/archive/refs/heads/main.zip"
-        commit_sha = ""
-        estimated_total = 0
-        estimated = False
-        estimated_note = ""
-        fallback_total = 5 * 1024 * 1024 * 1024  # 5 GiB
-
-        def set_job_progress(**state):
-            _update_background_job(job_id, status="running", **state)
-
-        try:
-            # Best-effort SHA for UI banner
-            try:
-                commit_info = requests.get(
-                    "https://api.github.com/repos/chazlarson/plex-test-libraries/commits/main",
-                    timeout=5,
-                ).json()
-                commit_sha = commit_info.get("sha", "")[:7]
-            except Exception:
-                commit_sha = ""
-
-            # Try to get total size first (lets UI show determination early)
-            total_size = 0
-            try:
-                head = requests.head(zip_url, allow_redirects=True, timeout=10)
-                total_size = int(head.headers.get("Content-Length", "0") or 0)
-            except Exception:
-                total_size = 0
-            if not total_size:
-                try:
-                    release_info = requests.get(
-                        "https://api.github.com/repos/chazlarson/plex-test-libraries/releases/latest",
-                        timeout=5,
-                    ).json()
-                    assets = release_info.get("assets") or []
-                    release_zip = next(
-                        (a for a in assets if str(a.get("name", "")).lower().endswith(".zip")),
-                        None,
-                    )
-                    if release_zip and int(release_zip.get("size", 0) or 0) > 0:
-                        estimated_total = int(release_zip.get("size", 0) or 0)
-                        total_size = estimated_total
-                        estimated = True
-                        estimated_note = "release"
-                except Exception:
-                    estimated_total = 0
-                    estimated = False
-            if not total_size:
-                try:
-                    repo_info = requests.get(
-                        "https://api.github.com/repos/chazlarson/plex-test-libraries",
-                        timeout=5,
-                    ).json()
-                    size_kb = int(repo_info.get("size", 0) or 0)
-                    if size_kb > 0:
-                        estimated_total = size_kb * 1024
-                        total_size = estimated_total
-                        estimated = True
-                        estimated_note = "repo"
-                except Exception:
-                    estimated_total = 0
-                    estimated = False
-            if not total_size or (estimated and total_size < fallback_total):
-                total_size = fallback_total
-                estimated = True
-                estimated_note = "fallback"
-
-            set_job_progress(
-                phase="download",
-                pct=0 if total_size else None,  # None => indeterminate until we know size
-                text="Downloading zip...",
-                downloaded=0,
-                total=total_size,
-                estimated=estimated,
-                estimated_note=estimated_note,
-            )
-
-            ok, msg = _ensure_rw_dir(tmp_root)
-            if not ok:
-                raise RuntimeError(msg)
-            # Clean only our own stale temp folders
-            try:
-                for entry in os.listdir(tmp_root):
-                    if entry.startswith("qs_test_libs_"):
-                        shutil.rmtree(os.path.join(tmp_root, entry), ignore_errors=True)
-            except Exception:
-                pass
-
-            with tempfile.TemporaryDirectory(prefix="qs_test_libs_", dir=tmp_root) as tmpdir:
-                zip_path = os.path.join(tmpdir, "main.zip")
-
-                # Stream download with throttled progress updates
-                downloaded = 0
-                last_push = 0.0
-                with requests.get(zip_url, stream=True, timeout=(30, 300)) as r:
-                    r.raise_for_status()
-
-                    # If HEAD failed, try to get size from GET
-                    if not total_size:
-                        try:
-                            total_size = int(r.headers.get("Content-Length", "0") or 0)
-                            if total_size:
-                                estimated = False
-                                estimated_note = ""
-                            set_job_progress(total=total_size, estimated=estimated, estimated_note=estimated_note)
-                        except Exception:
-                            total_size = 0
-
-                    chunk = 1024 * 1024  # 1 MiB
-                    with open(zip_path, "wb") as f:
-                        for part in r.iter_content(chunk_size=chunk):
-                            if not part:
-                                continue
-                            f.write(part)
-                            downloaded += len(part)
-
-                            now = time.time()
-                            if (now - last_push) > 0.5 or (total_size and downloaded >= total_size):
-                                pct = None
-                                if total_size:
-                                    pct = int(downloaded * 100 / total_size)
-                                set_job_progress(
-                                    phase="download",
-                                    pct=pct,
-                                    text="Downloading zip...",
-                                    downloaded=downloaded,
-                                    total=total_size,
-                                    estimated=estimated,
-                                    estimated_note=estimated_note,
-                                )
-                                last_push = now
-
-                # Extract with per-file progress
-                set_job_progress(
-                    phase="extract",
-                    pct=0,
-                    text="Extracting...",
-                    files_done=0,
-                    files_total=0,
-                )
-                with zipfile.ZipFile(zip_path, "r") as zip_ref:
-                    members = zip_ref.infolist()
-                    total_files = len(members) or 1
-                    files_done = 0
-                    last_push = 0.0
-
-                    for info in members:
-                        zip_ref.extract(info, tmpdir)
-                        files_done += 1
-
-                        now = time.time()
-                        if (now - last_push) > 0.2 or files_done == total_files:
-                            pct = int(files_done * 100 / total_files)
-                            set_job_progress(
-                                phase="extract",
-                                pct=pct,
-                                text=f"Extracting... {files_done}/{total_files} files",
-                                files_done=files_done,
-                                files_total=total_files,
-                            )
-                            last_push = now
-
-                extracted_dir = os.path.join(tmpdir, "plex-test-libraries-main")
-
-                # Finalize (replace folder)
-                set_job_progress(phase="finalize", pct=95, text="Finalizing...")
-                if os.path.exists(target_path):
-                    if not _safe_to_replace_test_libraries(target_path):
-                        raise RuntimeError("Target path exists but does not look like test libraries. Choose an empty folder or one containing test libraries.")
-                    shutil.rmtree(target_path, onerror=helpers.handle_remove_readonly)
-                shutil.move(extracted_dir, target_path)
-
-                # Write version marker (best effort)
-                if commit_sha:
-                    try:
-                        with open(os.path.join(target_path, ".test_libraries_version"), "w") as f:
-                            f.write(commit_sha)
-                    except Exception as e:
-                        helpers.ts_log(f"Warning: Failed to write SHA version file: {e}", level="WARNING")
-
-                # Permissions for non-Windows
-                if platform.system() in ["Linux", "Darwin"]:
-                    subprocess.run(["chmod", "-R", "777", target_path], check=False)
-
-                _complete_background_job(
-                    job_id,
-                    phase="done",
-                    success=True,
-                    pct=100,
-                    text="Installed/updated successfully.",
-                    target_path=resolved_path,
-                )
-
-        except Exception as e:
-            _fail_background_job(
-                job_id,
-                e,
-                phase="error",
-                success=False,
-                pct=0,
-                text=f"Error: {str(e)}",
-            )
-
-    threading.Thread(target=worker, daemon=True).start()
-    return jsonify(success=True, job_id=job_id, started_at=job.get("started_epoch"))
-
-
 def _schedule_quickstart_imagemaid_run_marker(imagemaid_root, mode=None, config_name=None, timeout_seconds=20):
     root = Path(imagemaid_root)
     log_dir = root / "config" / "logs"
@@ -14086,154 +13202,6 @@ def _schedule_quickstart_imagemaid_run_marker(imagemaid_root, mode=None, config_
         _write_quickstart_imagemaid_run_marker(root, mode=mode, config_name=config_name)
 
     threading.Thread(target=worker, daemon=True).start()
-
-
-@app.route("/clone-test-libraries-progress", methods=["GET"])
-def clone_test_libraries_progress():
-    job_id = request.args.get("job_id", "")
-    info = _get_background_job(job_id)
-    if not info:
-        return jsonify(success=False, message="Unknown job_id"), 404
-
-    # avoid duplicate kwarg: remove job's 'success' if present
-    info_no_flag = dict(info)
-    info_no_flag.pop("success", None)
-
-    return jsonify(success=True, **info_no_flag)
-
-
-@app.route("/clone-test-libraries-active", methods=["GET"])
-def clone_test_libraries_active():
-    active = _get_active_background_job("test_library_install")
-    if not active:
-        return jsonify(success=True, active=False)
-
-    job_id = active.get("job_id")
-    info = _get_background_job(job_id) or {}
-    phase = info.get("phase")
-    if phase in ["done", "error"]:
-        _clear_active_background_job("test_library_install", job_id=job_id)
-        return jsonify(success=True, active=False)
-
-    return jsonify(
-        success=True,
-        active=True,
-        job_id=job_id,
-        started_at=info.get("started_epoch"),
-        progress=info,
-    )
-
-
-@app.route("/clone-test-libraries", methods=["POST"])
-def clone_test_libraries():
-    data = request.get_json(silent=True) or {}
-    quickstart_root = data.get("quickstart_root", "")
-
-    if not quickstart_root:
-        return jsonify(success=False, message="Quickstart root path not provided.")
-
-    _, target_path, tmp_root, _, _ = _resolve_test_libraries_paths(quickstart_root)
-
-    resolved_path = os.path.abspath(target_path)
-    if _paths_overlap(tmp_root, target_path):
-        return jsonify(success=False, message="Temp and final paths must be different and cannot be nested.")
-    if not _safe_to_replace_test_libraries(target_path):
-        return jsonify(
-            success=False,
-            message="Target path exists but does not look like test libraries. Choose an empty folder or one containing test libraries.",
-        )
-
-    try:
-        # If already exists
-        if os.path.exists(target_path) and _test_libraries_present(target_path):
-            return jsonify(success=True, message="Test libraries already present (ZIP install).", target_path=resolved_path)
-
-        # ZIP fallback if git not found or Download failed
-        zip_url = "https://github.com/chazlarson/plex-test-libraries/archive/refs/heads/main.zip"
-        commit_sha = None
-        try:
-            commit_info = requests.get("https://api.github.com/repos/chazlarson/plex-test-libraries/commits/main", timeout=5).json()
-            commit_sha = commit_info.get("sha", "")[:7]
-        except Exception:
-            commit_sha = None
-
-        ok, msg = _ensure_rw_dir(tmp_root)
-        if not ok:
-            return jsonify(success=False, message=msg)
-        try:
-            for entry in os.listdir(tmp_root):
-                if entry.startswith("qs_test_libs_"):
-                    shutil.rmtree(os.path.join(tmp_root, entry), ignore_errors=True)
-        except Exception:
-            pass
-
-        with tempfile.TemporaryDirectory(prefix="qs_test_libs_", dir=tmp_root) as tmpdir:
-            zip_path = os.path.join(tmpdir, "main.zip")
-
-            with requests.get(zip_url, stream=True, timeout=(30, 300)) as r:
-                if r.status_code != 200:
-                    return jsonify(success=False, message="Failed to download ZIP fallback from GitHub.")
-                with open(zip_path, "wb") as f:
-                    for chunk in r.iter_content(chunk_size=1024 * 1024):
-                        if chunk:
-                            f.write(chunk)
-
-            with zipfile.ZipFile(zip_path, "r") as zip_ref:
-                zip_ref.extractall(tmpdir)
-
-            extracted_dir = os.path.join(tmpdir, "plex-test-libraries-main")
-            if os.path.exists(target_path):
-                if not _safe_to_replace_test_libraries(target_path):
-                    return jsonify(
-                        success=False,
-                        message="Target path exists but does not look like test libraries. Choose an empty folder or one containing test libraries.",
-                    )
-                shutil.rmtree(target_path, onerror=helpers.handle_remove_readonly)
-            shutil.move(extracted_dir, target_path)
-
-            if commit_sha:
-                try:
-                    with open(os.path.join(target_path, ".test_libraries_version"), "w") as f:
-                        f.write(commit_sha)
-                except Exception as e:
-                    helpers.ts_log(f"Warning: Failed to write SHA version file: {e}", level="WARNING")
-
-        if platform.system() in ["Linux", "Darwin"]:
-            subprocess.run(["chmod", "-R", "777", target_path], check=False)
-
-        return jsonify(success=True, message="Test libraries installed successfully.", target_path=resolved_path)
-
-    except Exception as e:
-        helpers.ts_log(f"Test library install failed: {e}", level="ERROR")
-        return jsonify(success=False, message="Unexpected error.")
-
-
-@app.route("/purge-test-libraries", methods=["POST"])
-def purge_test_libraries():
-    data = request.get_json(silent=True) or {}
-    quickstart_root = data.get("quickstart_root", "")
-
-    if not quickstart_root:
-        return jsonify(success=False, message="Quickstart root path not provided.")
-
-    _, target_path, _, _, _ = _resolve_test_libraries_paths(quickstart_root)
-
-    resolved_path = os.path.abspath(target_path)
-
-    try:
-        if not os.path.exists(resolved_path):
-            return jsonify(success=False, message="Test libraries folder does not exist.")
-        if not _test_libraries_present(resolved_path):
-            return jsonify(
-                success=False,
-                message="Target path does not look like test libraries. Refusing to delete.",
-            )
-
-        shutil.rmtree(resolved_path, onerror=helpers.handle_remove_readonly)
-        return jsonify(success=True, message=f"Test libraries deleted at: {resolved_path}")
-
-    except Exception as e:
-        return jsonify(success=False, message=f"Failed to delete folder:\n{str(e)}")
 
 
 @app.route("/validate_metadata_file", methods=["POST"])
